+++ examples/resources/metabase_permissions_group/import.sh
@@ -1,0 +1,1 @@
+$ terraform import metabase_permissions_group.example 1