+++ examples/resources/metabase_permissions_group/resource.tf
@@ -1,0 +1,3 @@
+resource "metabase_permissions_group" "example" {
+  name = "Example Group"
+}