+++ internal/provider/provider.go
@@ -107,6 +107,9 @@ func (p *MetabaseProvider) Configure(ctx context.Context, req provider.Configure
 func (p *MetabaseProvider) Resources(ctx context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
 		func() resource.Resource {
+			return &PermissionsGroupResource{provider: p}
+		},
+		func() resource.Resource {
 			return &UserResource{provider: p}
 		},
 	}